Parse arrangements properly in `arrangement` mode
There was an `if` statement in `_create_grammar_str_from_dict()` to
ensure `arrangement` would count as a start symbol. However, when
`string_type` was set to `arrangement`, the line could never be run
because the `if` statement was only executed when `string_type ==
message`. I modified the condition to run for either of these values of
`string_type`.