Maintain order of symbols when merging grammars
The symbols are read into sets, which removes the original ordering
present in the grammars. The order of symbols matters when parsing, so
this would cause parsing to sometimes return different results or fail
entirely on the same input. Re-sorting the symbols after the set
operations prevents the failures and makes the output consistent across
runs.