Make grammar creation deterministic (#89)
* Remove unused function in `grammar_utils.py`

* Fix type annotation for `_sort_grammar_keys()`

* Remove internal use of `allow_just_arrangement`

The docstrings say the parameter is deprecated, so I have removed it
everywhere but external interfaces.

* Parse arrangements properly in `arrangement` mode

There was an `if` statement in `_create_grammar_str_from_dict()` to
ensure `arrangement` would count as a start symbol. However, when
`string_type` was set to `arrangement`, the line could never be run
because the `if` statement was only executed when `string_type ==
message`. I modified the condition to run for either of these values of
`string_type`.

* Maintain order of symbols when merging grammars

The symbols are read into sets, which removes the original ordering
present in the grammars. The order of symbols matters when parsing, so
this would cause parsing to sometimes return different results or fail
entirely on the same input. Re-sorting the symbols after the set
operations prevents the failures and makes the output consistent across
runs.