Add `.editorconfig` file
This will make matching the style when not using VSCode easier. I used
the existing style in the repository to create it. My goal is to
maintain the style, not change it.

I used the third-party utility `editorconfig-checker` to find violations
of the newly created file and manually fixed them.