Validate argument lengths for `dataclass`s
They are required in the grammar and spec, so they should be enforced in
the `dataclass`s as well.

I originally implemented checks to require at least 2 powers for a PCE
and at least one sea for a CTO because those affected me personally, but
I figured I should fix all instances while I was at it.