Fix type of `SUP.province_no_coast`
The grammar, the spec, and the field's type annotation all say that the
field's value should be a province without its coast, but the object
constructed through parsing had a `Location` in the field instead.

I have fixed `DAIDEVisitor` to construct the object properly, added code
to fix an incorrectly passed-in type (like some other objects do), and
added tests to ensure there are no regressions.